+++ R/likelihood.R
@@ -7,17 +7,29 @@
 #'
 #' Computes the log-likelihood of the scaled logit model at a given set of
 #' parameter estimates (or the MLE if \code{pars} is not supplied).
+#' Either \code{fit} or \code{x}, \code{y} and \code{pars} need to be supplied.
 #'
 #' @param fit An object returned by \code{\link{sclr}}. Or a list with
 #'   parameters, x and y entries corresponding to the parameter matrix, 
 #'   model matrix and model response.
+#' @param x Model matrix. Will be taken from \code{fit} if \code{NULL}.
+#' @param y Model response. Will be taken from \code{fit} if \code{NULL}.
 #' @param pars A named vector of parameter values. If \code{NULL} then the
 #'   estimates from \code{fit} will be used.
 #'
 #' @export
-sclr_log_likelihood <- function(fit, pars = NULL) {
+sclr_log_likelihood <- function(fit = NULL, x = NULL, y = NULL, pars = NULL) {
   
-  if (is.null(pars)) pars <- fit$parameters
+  if (is.null(fit)) {
+    if (is.null(x) | is.null(y) | is.null(pars))
+      abort("likelihood requires fit or x, y and pars")
+  } else {
+    if (!is_sclr(fit))
+      abort("fit must be of type sclr")
+    if (is.null(x)) x <- model.matrix(fit)
+    if (is.null(y)) y <- model.response(model.frame(fit))
+    if (is.null(pars)) pars <- coef(fit)
+  }
   
   if (!is.matrix(pars)) {
     pars_vec <- pars
@@ -27,10 +39,9 @@ sclr_log_likelihood <- function(fit, pars = NULL) {
 
   lambda <- pars[1, ]
 
-  y <- fit$y
-  x <- fit$x
   exp_Xb <- get_exp_Xb(y, x, pars)
   
   l <- y * log(lambda) + (1 - y) * log(1 + exp_Xb - lambda) - log(1 + exp_Xb)
+  
   sum(l)
 }