+++ tests/testthat/test-likelihood.R
@@ -8,8 +8,12 @@ library(sclr)
 test_that("Likelihood works", {
   fit <- sclr(status ~ logHI, one_titre_data)
   expect_true(is.numeric(sclr_log_likelihood(fit)))
-  sclr_log_likelihood(fit, c("lambda" = 0.5, "beta_0" = 1, "beta_1" = 1))
+  sclr_log_likelihood(fit, pars = c("lambda" = 0.5, "beta_0" = 1, "beta_1" = 1))
   expect_true(is.numeric(sclr_log_likelihood(
-    fit, c("lambda" = 0.5, "beta_0" = 1, "beta_1" = 1)
+    fit, pars = c("lambda" = 0.5, "beta_0" = 1, "beta_1" = 1)
   )))
+  expect_error(sclr_log_likelihood())
+  expect_error(
+    sclr_log_likelihood(list(x = fit$x, y = fit$y), pars = coef(fit))
+  )
 })